fixing tests, and adding TODOs where we need them
